Add pilot.status.elasticsearch.version field, remove nodepool package

+++ pkg/apis/navigator/validation/elasticsearch.go
@@ -106,8 +106,8 @@ func ValidateElasticsearchCluster(esc *navigator.ElasticsearchCluster) field.Err
 	return allErrs
 }
 
-func countElasticsearchMasters(pools []navigator.ElasticsearchClusterNodePool) int64 {
-	masters := int64(0)
+func countElasticsearchMasters(pools []navigator.ElasticsearchClusterNodePool) int32 {
+	masters := int32(0)
 	for _, pool := range pools {
 		if containsElasticsearchRole(pool.Roles, navigator.ElasticsearchRoleMaster) {
 			masters += pool.Replicas
